Tuan 4

+++ src/main/java/com/tdtu/finalproject/UserRepository.java
@@ -5,8 +5,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 
     @Query("SELECT user FROM User user WHERE user.username = :username")
     public User getUserByUsername(@Param("username") String username);
+
+
 }
